feat(qtensor): add matmul overload
The torch.Tensor implementation seems to deal with non-contiguous inputs
just fine, but if we don't overload it for QTensor and let instead the
chain of aten operations flow, it starts with a sequence of expand/view
calls that are not compatible with non-contiguous inputs.
It is better to overload matmul anyway, as we might be able to call directly
an optimized kernel here.