feat(qlinear): do not quantize biases
Quantizing the biases does not make much sense in terms of memory
savings, especially for float16 as they need to be quantized to int16.
This is also not interesting in terms of compute, as if the QLinear
needs to return a QTensor, the int32 data must always be downscaled to
int8, which implies a dequantization.