feat(tensors): safe QModuleMixin serialization
Flattened tensors meta is serialized using only strings, which are
compatible with Abstract Syntax Trees evluation for converting back
to python generic types.
The only exception is qtype, for which we use a custom helper.
This makes QModuleMixin serialization compatible with the weights_only
parameter of torch.load.
Note that this does NOT make QTensor compatible with safe serialization,
as we don't override the pickle code used when directly saving QTensor
and it uses its own representation of flattened tensors using non-primitive
types.