fix(awq): disable AWQ kernel if output features < 128
The GEMM kernel crashes if output features are less than 128. It is
unclear if the issue comes from:
- the kernel itself,
- a wrong packing/formatting of the inputs (padding might be required),
- a wrong evaluation of the kernel parameters.
As a workaround, both GEMM and GEMV kernels are disabled if the number
of output features is less than 128.
Eventually, only the unpacking code should remain in quanto, and the
AWQ kernels should be replaced by the latest MARLIN kernels that are
better maintained.